==> databricks/mysql/precentation/1.types_of_joins.ipynb <==
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('basic').getOrCreate()

from pyspark.sql.functions import col

circuit_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/circuits/circuits.csv')


race_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/race/race.csv')


#### Inner Join

circuit_race_join_df = circuit_df.join(race_df,circuit_df.circuit_id == race_df.circuiit_id,'inner').select(circuit_df.name.alias('circuit_name'),circuit_df.location,circuit_df.country,race_df.name,race_df.round)


#### Left Join

circuit_race_join_df = circuit_df.join(race_df,circuit_df.circuit_id == race_df.circuiit_id,'left').select(circuit_df.name.alias('circuit_name'),circuit_df.location,circuit_df.country,race_df.name,race_df.round)


#### Right Join

circuit_race_join_df = circuit_df.join(race_df,circuit_df.circuit_id == race_df.circuiit_id,'right').select(circuit_df.name.alias('circuit_name'),circuit_df.location,circuit_df.country,race_df.name,race_df.round)


#### Full Join

circuit_race_join_df = circuit_df.join(race_df,circuit_df.circuit_id == race_df.circuiit_id,'full').select(circuit_df.name.alias('circuit_name'),circuit_df.location,circuit_df.country,race_df.name,race_df.round)


#### Semi Join

circuit_race_join_df = circuit_df.join(race_df,circuit_df.circuit_id == race_df.circuiit_id,'semi').select(circuit_df.name.alias('circuit_name'),circuit_df.location,circuit_df.country,race_df.name,race_df.round)


#### Anti Join


circuit_race_join_df = circuit_df.join(race_df,circuit_df.circuit_id == race_df.circuiit_id,'anti').select(circuit_df.name.alias('circuit_name'),circuit_df.location,circuit_df.country,race_df.name,race_df.round)
