==> databricks/mysql/precentation/2.race_result_aggregate.ipynb <==
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('basic').getOrCreate()

from pyspark.sql.functions import col,current_timestamp,count,countDistinct,sum,aggregate,desc,rank
from pyspark.sql.window import Window

races_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/race/race.csv')\
    .filter('race_year in (2019,2020)')\
    .withColumnRenamed('name', 'race_name')\
    .withColumnRenamed('race_timestamp', 'race_date')\
    .select('circuiit_id','race_id',col('race_year'),col('race_name'),col('race_date'))


circuits_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/circuits/circuits.csv')\
        .select('circuit_id',col('location').alias('circuit_location'))


drivers_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/drivers/drivers.csv')\
    .withColumnRenamed('name','driver_name')\
    .withColumnRenamed('number','driver_number')\
    .withColumnRenamed('nationality','driver_nationality')\
    .select('driver_id','driver_name','driver_number','driver_nationality')


constructors_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/constructors/constructors.csv')\
    .select('constructor_id',col('name').alias('team'))


results_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/result/result.csv')\
    .select('race_id','result_id','driver_id','constructor_id','grid','fastest_lap',col('time').alias('race_time'),'points','position')


race_join = circuits_df.join(races_df, circuits_df.circuit_id == races_df.circuiit_id, 'inner')\
    .join(results_df,races_df.race_id == results_df.race_id,'inner')\
    .join(drivers_df,drivers_df.driver_id == results_df.driver_id, 'inner')\
    .join(constructors_df,constructors_df.constructor_id == results_df.constructor_id,'inner')


df1 = race_join .select('race_year','race_name','race_date','circuit_location','driver_name','driver_number','driver_nationality','team','grid','fastest_lap','race_time','points','position')\
    .withColumn('current_date',current_timestamp())\
    .orderBy(race_join.points.desc())


df1.coalesce(1).write.option('header',True).format('csv').save('../tbls/presentation/race_results_twoyears')

race_result_agg_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/presentation/race_results_twoyears/race_results_twoyears.csv')\
    .filter(col('race_year') == 2020)


race_result_agg_df.select(count('*')).show()

race_result_agg_df.select(countDistinct('race_name')).show()

df1.where('driver_name="Lewis Hamilton"').select(sum('points'), countDistinct('race_name')).show()

race_result_agg_df.groupBy('circuit_location','driver_name')\
    .agg(sum('points').alias('sum_points'),countDistinct('race_name').alias('cnt_distinct_race_name'))\
    .orderBy('circuit_location').show(50)

rnk_df = df1.groupBy('race_year','driver_name')\
    .agg(sum('points').alias('total_points'),countDistinct('race_name').alias('number_of_races'))\
    .orderBy('race_year')


window_spec = Window.partitionBy('race_year').orderBy(desc('total_points'))

rnk_df.withColumn('rank',rank().over(window_spec)).show(50)
