==> databricks/mysql/precentation/3.race_results_driver_standing_agg_rank.ipynb <==
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('basic').getOrCreate()

from pyspark.sql.functions import col,current_timestamp,count,countDistinct,sum,aggregate,desc,rank,when
from pyspark.sql.window import Window

race_result_agg_df = spark.read.option('header',True).option('inferSchema',True).format('csv').load('../tbls/presentation/race_results_twoyears/race_results_twoyears.csv')\
    .filter('race_year=2020')


driver_standing_df = race_result_agg_df.groupBy('race_year','driver_name','driver_nationality','team')\
    .agg(sum('points').alias('total_points'),\
    count(when(col('position') ==1,True)).alias('wins'))

window_spec = Window.partitionBy('race_year')\
    .orderBy(desc('total_points'),desc('wins'))
driverstanding = driver_standing_df.withColumn('rank',rank().over(window_spec))

driverstanding.write.option('header',True).mode('overwrite').format('csv').save('../tbls/presentation/driverstanding/driverstanding.csv')
