==> databricks/mysql/precentation/7.create_table_using_pyspark_with_specificed_db_table_file_stored_in_dblocatiion.ipynb <==
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('basic').getOrCreate()

spark = SparkSession.builder \
    .appName("PullDataFromNonDefaultDB") \
    .config("spark.sql.warehouse.dir", "../tbls/dbs")\
    .enableHiveSupport() \
    .getOrCreate()

spark.sql("CREATE DATABASE IF NOT EXISTS process_db")


spark.sql('show databases').show()

spark.sql('use process_db')
spark.sql('show tables').show()

#### 1 Circuit table

circuit_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/circuits/circuits.csv')


circuit_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.circuits')


spark.sql('select * from process_db.circuits').show()

#### 2 Race table

result_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/result/result.csv')


result_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.race')

spark.sql('select * from process_db.race').show()

#### 3 Constructors table

constructors_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/constructors/constructors.csv')


constructors_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.constructors')

spark.sql('select * from process_db.constructors').show()

#### 4 Drivers table

drivers_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/drivers/drivers.csv')


drivers_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.drivers')

spark.sql('select * from process_db.drivers').show()

#### 5 Results table

results_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/result/result.csv')


results_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.results')

spark.sql('select * from process_db.results').show()

#### 6 PitStops table

pitstops_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/pitstops/pitstops.csv')


pitstops_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.pitstops')

spark.sql('select * from process_db.pitstops').show()

#### 7 LapTimes table

laptimes_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/laptimes/laptimes.csv')


laptimes_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.laptimes')

spark.sql('select * from process_db.laptimes').show()

#### 8 Qualifying table

qualifying_processed_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/qualifying/qualifying.csv')


qualifying_processed_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.qualifying')

spark.sql('select * from process_db.qualifying').show()

#### 1 Race Results table

race_results_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/presentation/race_results/race_result.csv')

race_results_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.raceresult')

spark.sql('select * from process_db.raceresult').show()

#### 2 Race Result with Two Years data table

race_results_twoyears_df = out = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/presentation/race_results_twoyears/race_results_twoyears.csv')

race_results_twoyears_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.race_results_twoyears')

spark.sql('select * from process_db.race_results_twoyears').show()

#### 3 Driver Standing table

driverstanding_df = spark.read.option('header',True).option('inferSchema',True).csv('../tbls/presentation/driverstanding/driverstanding.csv')


driverstanding_df.write.mode('overwrite').format('parquet').saveAsTable('process_db.driverstandiing')


spark.sql('select * from process_db.driverstandiing').show()
